// src/common-pages/proxy/proxy.type.ts
export interface AgentRuleLevelItem {
  level: number;
  teamBetting: string;
  teamMember: number;
  teamRecharge: string;
}

export interface CommissionItem {
  commissionLevel: number;
  tier1: string;
  tier2: string;
  tier3: string;
  tier4: string;
  tier1Max: string;
  tier1Min: string;
  tier2Max: string;
  tier2Min: string;
  tier3Max: string;
  tier3Min: string;
  tier4Max: string;
  tier4Min: string;
}

export interface AgentRuleData {
  agentLevelList: AgentRuleLevelItem[];
  casinoCommissionList: CommissionItem[];
  lotteryCommissionList: CommissionItem[];
  refundList: CommissionItem[];
  scratchCommissionList: CommissionItem[];
  sportCommissionList: CommissionItem[];
}